feat(tx): fetch utxo with pruning vin and add vout in tx pool

+++ core/const.go
@@ -34,7 +34,7 @@ const (
 	FixedGasPrice = uint64(50)
 
 	TransferFee         = FixedGasPrice * TransferGasLimit
-	InOutNumPerExtraFee = uint64(16)
+	InOutNumPerExtraFee = 16
 
 	SyncFlag = "sync"
 )
